refactor: move reports to be under User path
