+++ app/Filament/Resources/User/Report/ReportStep/Pages/ListReportSteps.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace App\Filament\Resources\Admin\Report\Pages;
+namespace App\Filament\Resources\User\Report\ReportStep\Pages;
 
-use App\Filament\Resources\Base\BaseViewResource;
-use App\Filament\Resources\Admin\Report;
+use App\Filament\Resources\User\Report\ReportStep;
+use App\Filament\Resources\Base\BaseListResources;
 
 /**
- * Class ViewReport.
+ * Class ListReportSteps.
  */
-class ViewReport extends BaseViewResource
+class ListReportSteps extends BaseListResources
 {
-    protected static string $resource = Report::class;
+    protected static string $resource = ReportStep::class;
 
     /**
      * Get the header actions available.
