+++ database/seeders/Auth/Permission/PermissionSeeder.php
@@ -11,7 +11,7 @@ use App\Models\Admin\Announcement;
 use App\Models\Admin\Dump;
 use App\Models\Admin\Feature;
 use App\Models\Admin\FeaturedTheme;
-use App\Models\Admin\Report;
+use App\Models\User\Report;
 use App\Models\Auth\Permission;
 use App\Models\Auth\Role;
 use App\Models\Auth\User;
